test(host): split opencode live session suites
