+++ packages/host/src/adapters/agent-sessions/opencode-live-session-adapter.test-support.ts
@@ -1,0 +1,212 @@
+import type {
+  OpencodeNativeApprovalReply,
+  OpencodeNativeQuestionReply,
+  OpencodeRuntimeSnapshotSource,
+  OpencodeSessionRuntimeConnection,
+  OpencodeSessionRuntimeSignal,
+  PrepareOpencodeSessionRuntime,
+} from "@openducktor/adapters-opencode-sdk";
+import type { RuntimeInstanceSummary } from "@openducktor/contracts";
+import { RUNTIME_DESCRIPTORS_BY_KIND } from "@openducktor/contracts";
+import { Effect } from "effect";
+import type { AgentSessionLiveAdapterChange } from "../../ports/agent-session-live-adapter-port";
+import type { RuntimeLiveSessionLifecyclePort } from "../../ports/runtime-live-session-lifecycle-port";
+
+export const runtime: RuntimeInstanceSummary = {
+  kind: "opencode",
+  runtimeId: "runtime-1",
+  repoPath: "/repo",
+  taskId: null,
+  role: "workspace",
+  workingDirectory: "/repo",
+  runtimeRoute: { type: "local_http", endpoint: "http://127.0.0.1:43123" },
+  startedAt: "2026-07-16T10:00:00.000Z",
+  descriptor: RUNTIME_DESCRIPTORS_BY_KIND.opencode,
+};
+
+export const ref = {
+  repoPath: "/repo",
+  runtimeKind: "opencode" as const,
+  workingDirectory: "/repo/worktree",
+  externalSessionId: "session-1",
+};
+
+export const nativeSource = (
+  overrides: Partial<OpencodeRuntimeSnapshotSource> = {},
+): OpencodeRuntimeSnapshotSource => ({
+  externalSessionId: "session-1",
+  sessionAssociation: { kind: "unbound" },
+  title: "Live OpenCode session",
+  workingDirectory: "/repo/worktree",
+  startedAt: "2026-07-16T10:01:00.000Z",
+  runtimeActivity: "idle",
+  pendingApprovals: [
+    {
+      requestId: "permission-1",
+      requestInstanceId: "native-permission-instance",
+      requestType: "file_change",
+      title: "Edit a file",
+      metadata: { source: "opencode" },
+    },
+  ],
+  pendingQuestions: [
+    {
+      requestId: "question-1",
+      requestInstanceId: "native-question-instance",
+      questions: [
+        {
+          header: "Confirm",
+          question: "Continue?",
+          options: [{ label: "Yes", description: "Continue" }],
+        },
+      ],
+    },
+  ],
+  ...overrides,
+});
+
+export const controlSummary = {
+  externalSessionId: "controlled-session",
+  runtimeKind: "opencode" as const,
+  workingDirectory: "/repo/worktree",
+  title: "Controlled session",
+  sessionAssociation: { kind: "workflow", taskId: "task-1", role: "build" } as const,
+  startedAt: "2026-07-16T10:02:00.000Z",
+  status: "running" as const,
+};
+
+type RuntimeHarness = {
+  readonly prepareRuntime: PrepareOpencodeSessionRuntime;
+  readonly emit: (signal: OpencodeSessionRuntimeSignal) => Promise<void>;
+  readonly approvalReplies: OpencodeNativeApprovalReply[];
+  readonly questionReplies: OpencodeNativeQuestionReply[];
+  readonly controlCalls: Array<{ operation: string; input: unknown }>;
+  readonly releaseCalls: string[];
+  readonly contextLoadCalls: string[];
+  readonly setSources: (sources: OpencodeRuntimeSnapshotSource[]) => void;
+};
+
+export const createRuntimeHarness = (
+  options: {
+    readonly sendUserMessageBarrier?: Promise<void>;
+    readonly onSendUserMessage?: () => void;
+  } = {},
+): RuntimeHarness => {
+  let listener: ((signal: OpencodeSessionRuntimeSignal) => void | Promise<void>) | null = null;
+  const sources = [nativeSource()];
+  const approvalReplies: OpencodeNativeApprovalReply[] = [];
+  const questionReplies: OpencodeNativeQuestionReply[] = [];
+  const controlCalls: Array<{ operation: string; input: unknown }> = [];
+  const releaseCalls: string[] = [];
+  const contextLoadCalls: string[] = [];
+
+  const connection: OpencodeSessionRuntimeConnection = {
+    readSessionSources: async () => sources,
+    loadContextUsage: async (input) => {
+      contextLoadCalls.push(input.externalSessionId);
+      return {
+        totalTokens: 999,
+        model: { providerId: "openai", modelId: "gpt-5.1" },
+      };
+    },
+    replyApproval: async (input) => {
+      approvalReplies.push(input);
+    },
+    replyQuestion: async (input) => {
+      questionReplies.push(input);
+    },
+    startSession: async (input) => {
+      controlCalls.push({ operation: "start", input });
+      return controlSummary;
+    },
+    resumeSession: async (input) => {
+      controlCalls.push({ operation: "resume", input });
+      return controlSummary;
+    },
+    forkSession: async (input) => {
+      controlCalls.push({ operation: "fork", input });
+      return controlSummary;
+    },
+    sendUserMessage: async (input) => {
+      controlCalls.push({ operation: "send", input });
+      options.onSendUserMessage?.();
+      await options.sendUserMessageBarrier;
+      return {
+        type: "user_message",
+        externalSessionId: input.externalSessionId,
+        timestamp: "2026-07-16T10:03:00.000Z",
+        messageId: "user-1",
+        message: "Hello",
+        parts: [{ kind: "text", text: "Hello" }],
+        state: "queued",
+      };
+    },
+    updateSessionModel: async (input) => {
+      controlCalls.push({ operation: "model", input });
+    },
+    stopSession: async (input) => {
+      controlCalls.push({ operation: "stop", input });
+    },
+    releaseSession: async (input) => {
+      controlCalls.push({ operation: "release", input });
+    },
+  };
+
+  return {
+    prepareRuntime: async (input) => ({
+      connection,
+      initialSources: sources,
+      initialContextUsageBySessionId: new Map([
+        [
+          "session-1",
+          {
+            totalTokens: 321,
+            model: {
+              runtimeKind: "opencode",
+              providerId: "openai",
+              modelId: "gpt-5",
+              variant: "high",
+            },
+          },
+        ],
+      ]),
+      startForwarding: async (nextListener) => {
+        listener = nextListener;
+      },
+      release: async () => {
+        releaseCalls.push(input.runtimeId);
+        listener = null;
+      },
+    }),
+    emit: async (signal) => {
+      if (!listener) {
+        throw new Error("Forwarding has not started.");
+      }
+      await listener(signal);
+    },
+    approvalReplies,
+    questionReplies,
+    controlCalls,
+    releaseCalls,
+    contextLoadCalls,
+    setSources: (nextSources) => {
+      sources.splice(0, sources.length, ...nextSources);
+    },
+  };
+};
+
+export const createLifecycle = (
+  changes: AgentSessionLiveAdapterChange[],
+): RuntimeLiveSessionLifecyclePort => ({
+  registerRuntimeAdapter: () => Effect.void,
+  releaseRuntime: () => Effect.succeed([]),
+  runAdapterMutation: (mutation) =>
+    mutation.pipe(
+      Effect.tap((result) =>
+        Effect.sync(() => {
+          changes.push(...result.changes);
+        }),
+      ),
+      Effect.map((result) => result.value),
+    ),
+});
