+++ packages/host/src/adapters/agent-sessions/opencode-live-session-control-adapter.test.ts
@@ -1,0 +1,440 @@
+import { describe, expect, test } from "bun:test";
+import { Effect } from "effect";
+import type {
+  AgentSessionLiveAdapterChange,
+  AgentSessionRuntimeAdapterPort,
+} from "../../ports/agent-session-live-adapter-port";
+import type { RuntimeLiveSessionLifecyclePort } from "../../ports/runtime-live-session-lifecycle-port";
+import { createOpenCodeLiveSessionAdapterPreparer } from "./opencode-live-session-adapter";
+import {
+  controlSummary,
+  createLifecycle,
+  createRuntimeHarness,
+  nativeSource,
+  ref,
+  runtime,
+} from "./opencode-live-session-adapter.test-support";
+
+describe("OpenCode live session controls", () => {
+  test("delegates controls while the host projection remains the only session authority", async () => {
+    const harness = createRuntimeHarness();
+    const publishedChanges: AgentSessionLiveAdapterChange[] = [];
+    const prepared = await Effect.runPromise(
+      createOpenCodeLiveSessionAdapterPreparer({
+        liveSessionLifecycle: createLifecycle(publishedChanges),
+        prepareRuntime: harness.prepareRuntime,
+      })(runtime),
+    );
+    await Effect.runPromise(prepared.startForwarding());
+    const adapter = prepared.adapter as AgentSessionRuntimeAdapterPort;
+    const controlRef = { ...ref, externalSessionId: "controlled-session" };
+    const startInput = {
+      repoPath: "/repo",
+      runtimeKind: "opencode" as const,
+      workingDirectory: "/repo/worktree",
+      sessionScope: { kind: "workflow" as const, taskId: "task-1", role: "build" as const },
+      systemPrompt: "Build it",
+    };
+
+    await expect(Effect.runPromise(adapter.startSession(startInput))).resolves.toEqual(
+      controlSummary,
+    );
+    expect(adapter.matches(controlRef)).toBe(true);
+    await Effect.runPromise(
+      adapter.resumeSession({
+        ...controlRef,
+        sessionScope: startInput.sessionScope,
+      }),
+    );
+    await Effect.runPromise(
+      adapter.forkSession({
+        ...startInput,
+        parentExternalSessionId: "parent-1",
+      }),
+    );
+    const accepted = await Effect.runPromise(
+      adapter.sendUserMessage({
+        ...controlRef,
+        sessionScope: startInput.sessionScope,
+        parts: [{ kind: "text", text: "Hello" }],
+      }),
+    );
+    expect(accepted.type).toBe("user_message");
+    expect(publishedChanges.filter((change) => change.type === "transcript_event")).toEqual([
+      {
+        type: "transcript_event",
+        event: {
+          type: "user_message",
+          externalSessionId: "controlled-session",
+          timestamp: "2026-07-16T10:03:00.000Z",
+          messageId: "user-1",
+          message: "Hello",
+          parts: [{ kind: "text", text: "Hello" }],
+          state: "queued",
+          sessionRef: controlRef,
+        },
+      },
+    ]);
+
+    await Effect.runPromise(adapter.updateSessionModel({ ...controlRef, model: null }));
+    await Effect.runPromise(adapter.stopSession(controlRef));
+    expect(adapter.matches(controlRef)).toBe(false);
+    await Effect.runPromise(
+      adapter.resumeSession({
+        ...controlRef,
+        sessionScope: startInput.sessionScope,
+      }),
+    );
+    expect(adapter.matches(controlRef)).toBe(true);
+    await Effect.runPromise(adapter.releaseSession(controlRef));
+    expect(adapter.matches(controlRef)).toBe(false);
+
+    expect(harness.controlCalls.map((call) => call.operation)).toEqual([
+      "start",
+      "resume",
+      "fork",
+      "send",
+      "model",
+      "stop",
+      "resume",
+      "release",
+    ]);
+    expect(harness.controlCalls[0]?.input).toMatchObject({
+      runtimeKind: "opencode",
+      runtimePolicy: { kind: "opencode" },
+      sessionScope: startInput.sessionScope,
+    });
+    await expect(Effect.runPromise(adapter.releaseRuntime())).resolves.toEqual([ref]);
+    expect(harness.releaseCalls).toEqual(["runtime-1"]);
+  });
+
+  test("keeps concurrent session events current while a slash command send is pending", async () => {
+    let resolveSendStarted: () => void = () => undefined;
+    let releaseSend: () => void = () => undefined;
+    const sendStarted = new Promise<void>((resolve) => {
+      resolveSendStarted = resolve;
+    });
+    const sendBarrier = new Promise<void>((resolve) => {
+      releaseSend = resolve;
+    });
+    const harness = createRuntimeHarness({
+      sendUserMessageBarrier: sendBarrier,
+      onSendUserMessage: resolveSendStarted,
+    });
+    harness.setSources([
+      nativeSource({ pendingApprovals: [], pendingQuestions: [] }),
+      nativeSource({
+        externalSessionId: "session-2",
+        title: "Other OpenCode session",
+        pendingApprovals: [],
+        pendingQuestions: [],
+      }),
+    ]);
+    const publishedChanges: AgentSessionLiveAdapterChange[] = [];
+    const prepared = await Effect.runPromise(
+      createOpenCodeLiveSessionAdapterPreparer({
+        liveSessionLifecycle: createLifecycle(publishedChanges),
+        prepareRuntime: harness.prepareRuntime,
+      })(runtime),
+    );
+    await Effect.runPromise(prepared.startForwarding());
+    const adapter = prepared.adapter as AgentSessionRuntimeAdapterPort;
+    const sending = Effect.runPromise(
+      adapter.sendUserMessage({
+        ...ref,
+        sessionScope: { kind: "workflow", taskId: "task-1", role: "build" },
+        parts: [
+          {
+            kind: "slash_command",
+            command: { id: "review", trigger: "review", title: "review", hints: [] },
+          },
+        ],
+      }),
+    );
+    await sendStarted;
+
+    const forwarding = harness.emit({
+      type: "transcript_event",
+      externalSessionId: "session-2",
+      event: {
+        type: "session_status",
+        externalSessionId: "session-2",
+        timestamp: "2026-07-16T10:04:00.000Z",
+        status: { type: "busy", message: null },
+      },
+    });
+
+    try {
+      try {
+        const result = await Promise.race([
+          forwarding.then(() => "forwarded" as const),
+          new Promise<"blocked">((resolve) => setTimeout(() => resolve("blocked"), 500)),
+        ]);
+        expect(result).toBe("forwarded");
+        await expect(
+          Effect.runPromise(
+            adapter.readRetainedSnapshot({ ...ref, externalSessionId: "session-2" }),
+          ),
+        ).resolves.toMatchObject({ type: "live", session: { activity: "running" } });
+
+        await harness.emit({
+          type: "transcript_event",
+          externalSessionId: "session-1",
+          event: {
+            type: "user_message",
+            externalSessionId: "session-1",
+            timestamp: "2026-07-16T10:04:01.000Z",
+            messageId: "user-live",
+            message: "/review",
+            parts: [{ kind: "text", text: "/review" }],
+            state: "queued",
+          },
+        });
+        expect(publishedChanges).toContainEqual({
+          type: "transcript_event",
+          event: expect.objectContaining({
+            type: "user_message",
+            externalSessionId: "session-1",
+            messageId: "user-live",
+          }),
+        });
+        await harness.emit({
+          type: "transcript_event",
+          externalSessionId: "session-1",
+          event: {
+            type: "session_idle",
+            externalSessionId: "session-1",
+            timestamp: "2026-07-16T10:05:00.000Z",
+          },
+        });
+      } finally {
+        releaseSend();
+        await sending;
+        await forwarding;
+      }
+
+      await expect(Effect.runPromise(adapter.readRetainedSnapshot(ref))).resolves.toMatchObject({
+        type: "live",
+        session: { activity: "idle" },
+      });
+    } finally {
+      await Effect.runPromise(adapter.releaseRuntime());
+    }
+  });
+
+  for (const operation of ["start", "resume", "fork"] as const) {
+    test(`reports the latest runtime association after ${operation} invalidation and refresh`, async () => {
+      const harness = createRuntimeHarness();
+      const publishedChanges: AgentSessionLiveAdapterChange[] = [];
+      const prepared = await Effect.runPromise(
+        createOpenCodeLiveSessionAdapterPreparer({
+          liveSessionLifecycle: createLifecycle(publishedChanges),
+          prepareRuntime: harness.prepareRuntime,
+        })(runtime),
+      );
+      await Effect.runPromise(prepared.startForwarding());
+      const adapter = prepared.adapter as AgentSessionRuntimeAdapterPort;
+      const controlRef = { ...ref, externalSessionId: "controlled-session" };
+      const sessionScope = {
+        kind: "workflow" as const,
+        taskId: "task-1",
+        role: "build" as const,
+      };
+      const startInput = {
+        repoPath: "/repo",
+        runtimeKind: "opencode" as const,
+        workingDirectory: "/repo/worktree",
+        sessionScope,
+        systemPrompt: "Build it",
+      };
+
+      if (operation === "start") {
+        await Effect.runPromise(adapter.startSession(startInput));
+      } else if (operation === "resume") {
+        await Effect.runPromise(
+          adapter.resumeSession({
+            ...controlRef,
+            sessionScope,
+          }),
+        );
+      } else {
+        await Effect.runPromise(
+          adapter.forkSession({
+            ...startInput,
+            parentExternalSessionId: "parent-1",
+          }),
+        );
+      }
+
+      publishedChanges.length = 0;
+      harness.setSources([
+        nativeSource({
+          externalSessionId: "controlled-session",
+          sessionAssociation: { kind: "unbound" },
+          title: "Refreshed runtime session",
+          runtimeActivity: "idle",
+          pendingApprovals: [],
+          pendingQuestions: [],
+        }),
+      ]);
+      await harness.emit({ type: "sessions_invalidated" });
+
+      await expect(Effect.runPromise(adapter.listRetainedSnapshots("/repo"))).resolves.toEqual([
+        expect.objectContaining({
+          ref: controlRef,
+          sessionAssociation: { kind: "unbound" },
+        }),
+      ]);
+      await expect(
+        Effect.runPromise(adapter.readRetainedSnapshot(controlRef)),
+      ).resolves.toMatchObject({
+        type: "live",
+        session: {
+          ref: controlRef,
+          sessionAssociation: { kind: "unbound" },
+        },
+      });
+      expect(
+        publishedChanges.filter(
+          (change) =>
+            change.type === "session_upsert" &&
+            change.snapshot.ref.externalSessionId === "controlled-session",
+        ),
+      ).toEqual([
+        {
+          type: "session_upsert",
+          snapshot: expect.objectContaining({
+            ref: controlRef,
+            sessionAssociation: { kind: "unbound" },
+          }),
+        },
+      ]);
+    });
+  }
+
+  test("commits an authoritative refresh only inside the host lifecycle mutation", async () => {
+    const harness = createRuntimeHarness();
+    let enterMutation: () => void = () => undefined;
+    let releaseMutation: () => void = () => undefined;
+    const mutationEntered = new Promise<void>((resolve) => {
+      enterMutation = resolve;
+    });
+    const mutationBarrier = new Promise<void>((resolve) => {
+      releaseMutation = resolve;
+    });
+    const publishedChanges: AgentSessionLiveAdapterChange[] = [];
+    const lifecycle: RuntimeLiveSessionLifecyclePort = {
+      registerRuntimeAdapter: () => Effect.void,
+      releaseRuntime: () => Effect.succeed([]),
+      runAdapterMutation: (mutation) =>
+        Effect.gen(function* () {
+          yield* Effect.sync(enterMutation);
+          yield* Effect.promise(() => mutationBarrier);
+          const result = yield* mutation;
+          publishedChanges.push(...result.changes);
+          return result.value;
+        }),
+    };
+    const prepared = await Effect.runPromise(
+      createOpenCodeLiveSessionAdapterPreparer({
+        liveSessionLifecycle: lifecycle,
+        prepareRuntime: harness.prepareRuntime,
+      })(runtime),
+    );
+    await Effect.runPromise(prepared.startForwarding());
+    const adapter = prepared.adapter as AgentSessionRuntimeAdapterPort;
+    harness.setSources([
+      nativeSource({
+        runtimeActivity: "running",
+        pendingApprovals: [],
+        pendingQuestions: [],
+      }),
+    ]);
+    const forwarding = harness.emit({ type: "sessions_invalidated" });
+    await mutationEntered;
+
+    const beforeCommit = await Effect.runPromise(adapter.listRetainedSnapshots("/repo"));
+    expect(beforeCommit[0]?.activity).toBe("waiting_for_question");
+    releaseMutation();
+    await forwarding;
+
+    const afterCommit = await Effect.runPromise(adapter.listRetainedSnapshots("/repo"));
+    expect(afterCommit[0]?.activity).toBe("running");
+    expect(publishedChanges).toEqual([
+      {
+        type: "session_upsert",
+        snapshot: expect.objectContaining({ ref, activity: "running" }),
+      },
+    ]);
+  });
+
+  test("projects status events without replacing them from a stale runtime read", async () => {
+    const harness = createRuntimeHarness();
+    harness.setSources([
+      nativeSource({
+        runtimeActivity: "idle",
+        pendingApprovals: [],
+        pendingQuestions: [],
+      }),
+      nativeSource({
+        externalSessionId: "session-2",
+        title: "Other OpenCode session",
+        runtimeActivity: "idle",
+        pendingApprovals: [],
+        pendingQuestions: [],
+      }),
+    ]);
+    const publishedChanges: AgentSessionLiveAdapterChange[] = [];
+    const prepared = await Effect.runPromise(
+      createOpenCodeLiveSessionAdapterPreparer({
+        liveSessionLifecycle: createLifecycle(publishedChanges),
+        prepareRuntime: harness.prepareRuntime,
+      })(runtime),
+    );
+    await Effect.runPromise(prepared.startForwarding());
+    const adapter = prepared.adapter as AgentSessionRuntimeAdapterPort;
+
+    await harness.emit({
+      type: "transcript_event",
+      externalSessionId: "session-1",
+      event: {
+        type: "session_status",
+        externalSessionId: "session-1",
+        timestamp: "2026-07-16T10:02:00.000Z",
+        status: { type: "busy", message: null },
+      },
+    });
+
+    await expect(Effect.runPromise(adapter.readRetainedSnapshot(ref))).resolves.toMatchObject({
+      type: "live",
+      session: { activity: "running" },
+    });
+    await expect(
+      Effect.runPromise(adapter.readRetainedSnapshot({ ...ref, externalSessionId: "session-2" })),
+    ).resolves.toMatchObject({
+      type: "live",
+      session: { activity: "idle" },
+    });
+
+    await harness.emit({
+      type: "transcript_event",
+      externalSessionId: "session-1",
+      event: {
+        type: "session_status",
+        externalSessionId: "session-1",
+        timestamp: "2026-07-16T10:03:00.000Z",
+        status: { type: "idle" },
+      },
+    });
+
+    await expect(Effect.runPromise(adapter.readRetainedSnapshot(ref))).resolves.toMatchObject({
+      type: "live",
+      session: { activity: "idle" },
+    });
+    expect(
+      publishedChanges
+        .filter((change) => change.type === "session_upsert")
+        .map((change) => (change.type === "session_upsert" ? change.snapshot.activity : null)),
+    ).toEqual(["running", "idle"]);
+  });
+});
